+++ sdk/storage/Azure.Storage.Files.Shares/tests/ShareClientTests.cs
@@ -463,6 +463,39 @@ namespace Azure.Storage.Files.Shares.Test
         }
 
         [Test]
+        public async Task GetPropertiesAsync_Snapshot()
+        {
+            // Arrange
+            await using DisposingShare test = await GetTestShareAsync();
+            ShareClient share = test.Share;
+
+            Response<ShareSnapshotInfo> createSnapshotResponse = await share.CreateSnapshotAsync();
+            ShareClient snapshotShareClient = share.WithSnapshot(createSnapshotResponse.Value.Snapshot);
+
+            // Act
+            Response<ShareProperties> response = await snapshotShareClient.GetPropertiesAsync();
+
+            // Assert
+            Assert.IsNotNull(response.Value.ETag);
+            Assert.IsNotNull(response.Value.LastModified);
+        }
+
+        [Test]
+        public async Task GetPropertiesAsync_SnapshotFailed()
+        {
+            // Arrange
+            await using DisposingShare test = await GetTestShareAsync();
+            ShareClient share = test.Share;
+
+            ShareClient snapshotShareClient = share.WithSnapshot("2020-06-26T00:49:21.0000000Z");
+
+            // Act
+            await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
+                snapshotShareClient.GetPropertiesAsync(),
+                e => Assert.AreEqual(ShareErrorCode.ShareNotFound.ToString(), e.ErrorCode));
+        }
+
+        [Test]
         [Ignore("#10044: Re-enable failing Storage tests")]
         public async Task GetPropertiesAsync_Premium()
         {
@@ -721,6 +754,48 @@ namespace Azure.Storage.Files.Shares.Test
         }
 
         [Test]
+        public async Task DeleteAsync_Snapshot()
+        {
+            // Arrange
+            var shareName = GetNewShareName();
+            ShareServiceClient service = GetServiceClient_SharedKey();
+            ShareClient share = InstrumentClient(service.GetShareClient(shareName));
+
+            await share.CreateAsync(quotaInGB: 1);
+
+            Response<ShareSnapshotInfo> response = await share.CreateSnapshotAsync();
+            ShareClient snapshotShareClient = share.WithSnapshot(response.Value.Snapshot);
+
+            // Act
+            await snapshotShareClient.DeleteAsync(false);
+
+            // Assert
+            Response<bool> shareExistsResponse = await share.ExistsAsync();
+            Response<bool> snapshotExistsResponse = await snapshotShareClient.ExistsAsync();
+
+            Assert.IsTrue(shareExistsResponse.Value);
+            Assert.IsFalse(snapshotExistsResponse.Value);
+
+            // Cleanup
+            await share.DeleteAsync();
+        }
+
+        [Test]
+        public async Task DeleteAsync_SnapshotFailed()
+        {
+            // Arrange
+            await using DisposingShare test = await GetTestShareAsync();
+            ShareClient share = test.Share;
+
+            ShareClient snapshotShareClient = share.WithSnapshot("2020-06-26T00:49:21.0000000Z");
+
+            // Act
+            await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
+               snapshotShareClient.DeleteAsync(),
+               e => Assert.AreEqual(ShareErrorCode.InvalidQueryParameterValue.ToString(), e.ErrorCode));
+        }
+
+        [Test]
         public async Task Delete_Error()
         {
             // Arrange
@@ -732,7 +807,7 @@ namespace Azure.Storage.Files.Shares.Test
             // Act
             await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
                 share.DeleteAsync(false),
-                e => Assert.AreEqual("ShareNotFound", e.ErrorCode));
+                e => Assert.AreEqual(ShareErrorCode.ShareNotFound.ToString(), e.ErrorCode));
         }
 
         [Test]