+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added support for Large Files.
 - Added support for File Soft Delete.
 - Fixed bug where ShareDirectoryClient and ShareFileClient.Name and .Path were sometimes URL-encoded.
+- Fixed bug where ShareClient.WithSnapshot(), ShareDirectoryClient.WithSnapshot(), and ShareFileClient.WithSnapshot() were not functioning correctly.
 
 ## 12.2.3 (2020-06)
 - This release contains bug fixes to improve quality.